+++ kubernetes-model-generator/kubernetes-model-core/src/test/java/io/fabric8/kubernetes/api/model/ProjectedVolumeSourceTest.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright (C) 2015 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.fabric8.kubernetes.api.model;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.assertj.core.api.InstanceOfAssertFactories;
+import org.junit.jupiter.api.Test;
+
+import java.io.IOException;
+import java.io.InputStream;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+class ProjectedVolumeSourceTest {
+  private final ObjectMapper objectMapper = new ObjectMapper();
+
+  @Test
+  void deserializationWithOctalValueShouldWorkAsExpected() throws IOException {
+    // Given
+    InputStream inputStream = getClass().getResourceAsStream("/projectedvolumesource.json");
+
+    // When
+    ProjectedVolumeSource projectedVolumeSource = objectMapper.readValue(inputStream, ProjectedVolumeSource.class);
+
+    // Then
+    assertThat(projectedVolumeSource)
+        .hasFieldOrPropertyWithValue("defaultMode", Integer.valueOf("0555", 8))
+        .extracting(ProjectedVolumeSource::getSources)
+        .asInstanceOf(InstanceOfAssertFactories.list(VolumeProjection.class))
+        .singleElement()
+        .extracting(VolumeProjection::getSecret)
+        .hasFieldOrPropertyWithValue("name", "mysecret")
+        .extracting(SecretProjection::getItems)
+        .asInstanceOf(InstanceOfAssertFactories.list(KeyToPath.class))
+        .singleElement()
+        .hasFieldOrPropertyWithValue("key", "username")
+        .hasFieldOrPropertyWithValue("path", "my-group/my-username");
+  }
+}